pyproject.toml
[project]
name = "ancient-books"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "easyocr>=1.7.2",
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
]
